cleanup alica_viewer
