+++ alica_viewer/include/alica_viewer/alica_viewer_ros_interface.h
@@ -1,5 +1,4 @@
-#ifndef ALICA_VIEWER_ALICA_VIEWER_ROS_INTERFACE_H
-#define ALICA_VIEWER_ALICA_VIEWER_ROS_INTERFACE_H
+#pragma once
 
 #include <QThread>
 #include <ros/ros.h>
@@ -41,5 +40,3 @@ class AlicaViewerRosInterface : public QThread
 // Declare templates that can be accepted by QVariant
 Q_DECLARE_METATYPE(alica::AlicaEngineInfo);
 Q_DECLARE_METATYPE(alica::PlanTreeInfo);
-
-#endif // ALICA_VIEWER_ALICA_VIEWER_ROS_INTERFACE_H
